refactor: Replace utils.ExecuteConcurrently with concurrent.Run for task execution
